Move BotUtils to utility package
